feat(backend): remove applications related changes
