+++ backend/core/src/applications/application-preference-api-extra-models.ts
@@ -1,0 +1,5 @@
+import { BooleanInput } from "./types/form-metadata/boolean-input"
+import { TextInput } from "./types/form-metadata/text-input"
+import { AddressInput } from "./types/form-metadata/address-input"
+
+export const applicationPreferenceApiExtraModels = [BooleanInput, TextInput, AddressInput]
