+++ backend/core/src/applications/applications.controller.ts
@@ -13,24 +13,154 @@ import {
   UsePipes,
   ValidationPipe,
 } from "@nestjs/common"
-import { ApiBearerAuth, ApiExtraModels, ApiOperation, ApiTags } from "@nestjs/swagger"
+import { ApiBearerAuth, ApiExtraModels, ApiOperation, ApiProperty, ApiTags } from "@nestjs/swagger"
 import { OptionalAuthGuard } from "../auth/guards/optional-auth.guard"
 import { ResourceType } from "../auth/decorators/resource-type.decorator"
 import { mapTo } from "../shared/mapTo"
-import { ApplicationDto } from "./dto/application.dto"
+import {
+  ApplicationCreateDto,
+  ApplicationDto,
+  ApplicationUpdateDto,
+  PaginatedApplicationDto,
+} from "./dto/application.dto"
+import { Expose, Transform } from "class-transformer"
+import { IsBoolean, IsOptional, IsString, IsIn } from "class-validator"
+import { PaginationQueryParams } from "../shared/dto/pagination.dto"
 import { ValidationsGroupsEnum } from "../shared/types/validations-groups-enum"
 import { defaultValidationPipeOptions } from "../shared/default-validation-pipe-options"
-import { applicationPreferenceApiExtraModels } from "./types/application-preference-api-extra-models"
+import { applicationPreferenceApiExtraModels } from "./application-preference-api-extra-models"
 import { ListingsService } from "../listings/listings.service"
-import { ActivityLogInterceptor } from "../activity-log/interceptors/activity-log.interceptor"
 import { ApplicationCsvExporterService } from "./services/application-csv-exporter.service"
 import { ApplicationsService } from "./services/applications.service"
-import { PaginatedApplicationListQueryParams } from "./dto/paginated-application-list-query-params"
-import { ApplicationsCsvListQueryParams } from "./dto/applications-csv-list-query-params"
-import { ApplicationsApiExtraModel } from "./types/applications-api-extra-model"
-import { PaginatedApplicationDto } from "./dto/paginated-application.dto"
-import { ApplicationCreateDto } from "./dto/application-create.dto"
-import { ApplicationUpdateDto } from "./dto/application-update.dto"
+import { ActivityLogInterceptor } from "../activity-log/interceptors/activity-log.interceptor"
+
+export enum OrderByParam {
+  firstName = "applicant.firstName",
+  lastName = "applicant.lastName",
+  submissionDate = "application.submissionDate",
+  createdAt = "application.createdAt",
+}
+
+export enum OrderParam {
+  ASC = "ASC",
+  DESC = "DESC",
+}
+
+class ApplicationsApiExtraModel {
+  @Expose()
+  @ApiProperty({
+    enum: Object.keys(OrderByParam),
+    example: "createdAt",
+    default: "createdAt",
+    required: false,
+  })
+  orderBy?: OrderByParam
+
+  @Expose()
+  @ApiProperty({
+    enum: OrderParam,
+    example: "DESC",
+    default: "DESC",
+    required: false,
+  })
+  order?: OrderParam
+}
+
+export class PaginatedApplicationListQueryParams extends PaginationQueryParams {
+  @Expose()
+  @ApiProperty({
+    type: String,
+    example: "listingId",
+    required: false,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  listingId?: string
+
+  @Expose()
+  @ApiProperty({
+    type: String,
+    example: "search",
+    required: false,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  search?: string
+
+  @Expose()
+  @ApiProperty({
+    type: String,
+    example: "userId",
+    required: false,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  userId?: string
+
+  @Expose()
+  @ApiProperty({
+    enum: Object.keys(OrderByParam),
+    example: "createdAt",
+    default: "createdAt",
+    required: false,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @IsIn(Object.values(OrderByParam), { groups: [ValidationsGroupsEnum.default] })
+  @Transform((value: string | undefined) =>
+    value ? (OrderByParam[value] ? OrderByParam[value] : value) : OrderByParam.createdAt
+  )
+  orderBy?: OrderByParam
+
+  @Expose()
+  @ApiProperty({
+    enum: OrderParam,
+    example: "DESC",
+    default: "DESC",
+    required: false,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @IsIn(Object.keys(OrderParam), { groups: [ValidationsGroupsEnum.default] })
+  @Transform((value: string | undefined) => (value ? value : OrderParam.DESC))
+  order?: OrderParam
+
+  @Expose()
+  @ApiProperty({
+    type: Boolean,
+    example: true,
+    required: false,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
+  @Transform(
+    (value: string | undefined) => {
+      switch (value) {
+        case "true":
+          return true
+        case "false":
+          return false
+        default:
+          return undefined
+      }
+    },
+    { toClassOnly: true }
+  )
+  markedAsDuplicate?: boolean
+}
+
+export class ApplicationsCsvListQueryParams extends PaginatedApplicationListQueryParams {
+  @Expose()
+  @ApiProperty({
+    type: Boolean,
+    example: true,
+    required: false,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
+  @Transform((value: string | undefined) => value === "true", { toClassOnly: true })
+  includeDemographics?: boolean
+}
 
 @Controller("applications")
 @ApiTags("applications")
