+++ backend/core/src/applications/applications.module.ts
@@ -11,22 +11,24 @@ import { ApplicationsSubmissionController } from "./applications-submission.cont
 import { ApplicationFlaggedSetsModule } from "../application-flagged-sets/application-flagged-sets.module"
 import { TranslationsModule } from "../translations/translations.module"
 import { Listing } from "../listings/entities/listing.entity"
-import { ActivityLogModule } from "../activity-log/activity-log.module"
+import { ScheduleModule } from "@nestjs/schedule"
 import { ApplicationsService } from "./services/applications.service"
 import { CsvBuilder } from "./services/csv-builder.service"
 import { ApplicationCsvExporterService } from "./services/application-csv-exporter.service"
 import { EmailModule } from "../email/email.module"
+import { ActivityLogModule } from "../activity-log/activity-log.module"
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([Application, Applicant, Address, Listing]),
     AuthModule,
+    ActivityLogModule,
     SharedModule,
     ListingsModule,
     ApplicationFlaggedSetsModule,
     TranslationsModule,
     EmailModule,
-    ActivityLogModule,
+    ScheduleModule.forRoot(),
   ],
   providers: [ApplicationsService, CsvBuilder, ApplicationCsvExporterService],
   exports: [ApplicationsService],
