--- backend/core/src/applications/dto/application-create.dto.ts
@@ -1,86 +1,0 @@
-import { OmitType } from "@nestjs/swagger"
-import { Expose, Type } from "class-transformer"
-import { ArrayMaxSize, IsDefined, ValidateNested } from "class-validator"
-import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
-import { IdDto } from "../../shared/dto/id.dto"
-import { ApplicantCreateDto } from "./applicant.dto"
-import { AddressCreateDto } from "../../shared/dto/address.dto"
-import { AlternateContactCreateDto } from "./alternate-contact.dto"
-import { AccessibilityCreateDto } from "./accessibility.dto"
-import { DemographicsCreateDto } from "./demographics.dto"
-import { HouseholdMemberCreateDto } from "./household-member.dto"
-import { ApplicationDto } from "./application.dto"
-
-export class ApplicationCreateDto extends OmitType(ApplicationDto, [
-  "id",
-  "createdAt",
-  "updatedAt",
-  "deletedAt",
-  "applicant",
-  "listing",
-  "user",
-  "mailingAddress",
-  "alternateAddress",
-  "alternateContact",
-  "accessibility",
-  "demographics",
-  "householdMembers",
-  "markedAsDuplicate",
-  "preferredUnit",
-  "confirmationCode",
-] as const) {
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => IdDto)
-  listing: IdDto
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => ApplicantCreateDto)
-  applicant: ApplicantCreateDto
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressCreateDto)
-  mailingAddress: AddressCreateDto
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressCreateDto)
-  alternateAddress: AddressCreateDto
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AlternateContactCreateDto)
-  alternateContact: AlternateContactCreateDto
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AccessibilityCreateDto)
-  accessibility: AccessibilityCreateDto
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => DemographicsCreateDto)
-  demographics: DemographicsCreateDto
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @ArrayMaxSize(32, { groups: [ValidationsGroupsEnum.default] })
-  @Type(() => HouseholdMemberCreateDto)
-  householdMembers: HouseholdMemberCreateDto[]
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => IdDto)
-  preferredUnit: IdDto[]
-}
