+++ backend/core/src/applications/dto/application.dto.ts
@@ -1,14 +1,34 @@
 import { OmitType } from "@nestjs/swagger"
-import { ArrayMaxSize, IsDefined, ValidateNested } from "class-validator"
+import {
+  ArrayMaxSize,
+  IsDate,
+  IsDefined,
+  IsOptional,
+  IsUUID,
+  ValidateNested,
+} from "class-validator"
 import { Application } from "../entities/application.entity"
 import { Expose, plainToClass, Transform, Type } from "class-transformer"
 import { IdDto } from "../../shared/dto/id.dto"
-import { ApplicantDto } from "./applicant.dto"
-import { AddressDto } from "../../shared/dto/address.dto"
-import { AlternateContactDto } from "./alternate-contact.dto"
-import { DemographicsDto } from "./demographics.dto"
-import { HouseholdMemberDto } from "./household-member.dto"
-import { AccessibilityDto } from "./accessibility.dto"
+import { PaginationFactory } from "../../shared/dto/pagination.dto"
+import { ApplicantCreateDto, ApplicantDto, ApplicantUpdateDto } from "./applicant.dto"
+import { AddressCreateDto, AddressDto, AddressUpdateDto } from "../../shared/dto/address.dto"
+import {
+  AlternateContactCreateDto,
+  AlternateContactDto,
+  AlternateContactUpdateDto,
+} from "./alternate-contact.dto"
+import { DemographicsCreateDto, DemographicsDto, DemographicsUpdateDto } from "./demographics.dto"
+import {
+  HouseholdMemberCreateDto,
+  HouseholdMemberDto,
+  HouseholdMemberUpdateDto,
+} from "./household-member.dto"
+import {
+  AccessibilityCreateDto,
+  AccessibilityDto,
+  AccessibilityUpdateDto,
+} from "./accessibility.dto"
 import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
 import { UnitTypeDto } from "../../unit-types/dto/unit-type.dto"
 
@@ -98,4 +118,176 @@ export class ApplicationDto extends OmitType(Application, [
   @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
   @Type(() => UnitTypeDto)
   preferredUnit: UnitTypeDto[]
+}
+
+export class PaginatedApplicationDto extends PaginationFactory<ApplicationDto>(ApplicationDto) {}
+
+export class ApplicationCreateDto extends OmitType(ApplicationDto, [
+  "id",
+  "createdAt",
+  "updatedAt",
+  "deletedAt",
+  "applicant",
+  "listing",
+  "user",
+  "mailingAddress",
+  "alternateAddress",
+  "alternateContact",
+  "accessibility",
+  "demographics",
+  "householdMembers",
+  "markedAsDuplicate",
+  "preferredUnit",
+  "confirmationCode",
+] as const) {
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => IdDto)
+  listing: IdDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => ApplicantCreateDto)
+  applicant: ApplicantCreateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AddressCreateDto)
+  mailingAddress: AddressCreateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AddressCreateDto)
+  alternateAddress: AddressCreateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AlternateContactCreateDto)
+  alternateContact: AlternateContactCreateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AccessibilityCreateDto)
+  accessibility: AccessibilityCreateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => DemographicsCreateDto)
+  demographics: DemographicsCreateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
+  @ArrayMaxSize(32, { groups: [ValidationsGroupsEnum.default] })
+  @Type(() => HouseholdMemberCreateDto)
+  householdMembers: HouseholdMemberCreateDto[]
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
+  @Type(() => IdDto)
+  preferredUnit: IdDto[]
+}
+
+export class ApplicationUpdateDto extends OmitType(ApplicationDto, [
+  "id",
+  "createdAt",
+  "updatedAt",
+  "deletedAt",
+  "applicant",
+  "listing",
+  "user",
+  "mailingAddress",
+  "alternateAddress",
+  "alternateContact",
+  "accessibility",
+  "demographics",
+  "householdMembers",
+  "markedAsDuplicate",
+  "preferredUnit",
+  "confirmationCode",
+] as const) {
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsUUID(4, { groups: [ValidationsGroupsEnum.default] })
+  id?: string
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsDate({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => Date)
+  createdAt?: Date
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsDate({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => Date)
+  updatedAt?: Date
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsDate({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => Date)
+  deletedAt?: Date
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => IdDto)
+  listing: IdDto
+
+  @Expose()
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => ApplicantUpdateDto)
+  applicant: ApplicantUpdateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AddressUpdateDto)
+  mailingAddress: AddressUpdateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AddressUpdateDto)
+  alternateAddress: AddressUpdateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AlternateContactUpdateDto)
+  alternateContact: AlternateContactUpdateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => AccessibilityUpdateDto)
+  accessibility: AccessibilityUpdateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
+  @Type(() => DemographicsUpdateDto)
+  demographics: DemographicsUpdateDto
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
+  @ArrayMaxSize(32, { groups: [ValidationsGroupsEnum.default] })
+  @Type(() => HouseholdMemberUpdateDto)
+  householdMembers: HouseholdMemberUpdateDto[]
+
+  @Expose()
+  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
+  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
+  @Type(() => IdDto)
+  preferredUnit: IdDto[]
 }
