--- backend/core/src/applications/dto/paginated-application.dto.ts
@@ -1,4 +1,0 @@
-import { PaginationFactory } from "../../shared/dto/pagination.dto"
-import { ApplicationDto } from "./application.dto"
-
-export class PaginatedApplicationDto extends PaginationFactory<ApplicationDto>(ApplicationDto) {}
