--- backend/core/src/applications/types/application-preference-api-extra-models.ts
@@ -1,5 +1,0 @@
-import { BooleanInput } from "./form-metadata/boolean-input"
-import { TextInput } from "./form-metadata/text-input"
-import { AddressInput } from "./form-metadata/address-input"
-
-export const applicationPreferenceApiExtraModels = [BooleanInput, TextInput, AddressInput]
