--- backend/core/src/applications/types/order-by-param.ts
@@ -1,6 +1,0 @@
-export enum OrderByParam {
-  firstName = "applicant.firstName",
-  lastName = "applicant.lastName",
-  submissionDate = "application.submissionDate",
-  createdAt = "application.createdAt",
-}
