--- backend/core/src/applications/types/order-param.ts
@@ -1,4 +1,0 @@
-export enum OrderParam {
-  ASC = "ASC",
-  DESC = "DESC",
-}
